tests/unit/GameState.spec.ts
import GameState from "@/game/GameState";

test("getGameResult>>0", () => {
  const state = [
    [0, 0, 0, 0, 0],
    [0, 0, 0, 0, 0],
    [0, 0, 0, 0, 0],
    [0, 0, 0, 0, 0],
    [0, 0, 0, 0, 0]
  ];
  const gameState = new GameState(state, 1);
  expect(gameState.getGameResult()).toBe(0);
});

test("getGameResult>>1 横向", () => {
  const state = [
    [0, 0, 0, 0, 0],
    [0, 0, 0, 0, 0],
    [0, 0, 0, 0, 0],
    [0, 0, 0, 0, 0],
    [1, 1, 1, 1, 1]
  ];
  const gameState = new GameState(state, 1);
  expect(gameState.getGameResult()).toBe(1);
});

test("getGameResult>>1 横向", () => {
  const state = [
    [0, 0, 0, 0, 0, 0, 0, 0],
    [0, 0, 0, 0, 0, 0, 0, 0],
    [0, 0, 0, 0, 0, 0, 0, 0],
    [0, 0, 0, 0, 0, 0, 0, 0],
    [0, 1, 1, 1, 1, 1, 0, 0],
    [0, 0, 0, 0, 0, 0, 0, 0],
    [0, 0, 0, 0, 0, 0, 0, 0],
    [0, 0, 0, 0, 0, 0, 0, 0]
  ];
  const gameState = new GameState(state, 1);
  expect(gameState.getGameResult()).toBe(1);
});

test("getGameResult>>1 纵向", () => {
  const state = [
    [0, 0, 1, 0, 0, 0, 0, 0],
    [0, 0, 1, 0, 0, 0, 0, 0],
    [0, 0, 1, 0, 0, 0, 0, 0],
    [0, 0, 1, 0, 0, 0, 0, 0],
    [0, 1, 1, 1, 1, 0, 0, 0],
    [0, 0, 0, 0, 0, 0, 0, 0],
    [0, 0, 0, 0, 0, 0, 0, 0],
    [0, 0, 0, 0, 0, 0, 0, 0]
  ];
  const gameState = new GameState(state, 1);
  expect(gameState.getGameResult()).toBe(1);
});

test("getGameResult>>1 斜向", () => {
  const state = [
    [1, 0, 0, 0, 0],
    [0, 1, 0, 0, 0],
    [0, 0, 1, 0, 0],
    [0, 0, 0, 1, 0],
    [1, 1, 0, 1, 1]
  ];
  const gameState = new GameState(state, 1);
  expect(gameState.getGameResult()).toBe(1);
});
